# etcdutil, mcs: fix the issue loading label rules is too slow