<!--

Thank you for working on PD! Please read PD's [CONTRIBUTING](https://github.com/tikv/pd/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

<!--

Please create an issue first to describe the problem.
There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".
For more info, check https://github.com/tikv/pd/blob/master/CONTRIBUTING.md#linking-issues.

-->
Issue Number: Close #7724

### What is changed and how does it work?

<!--

You could use the "commit message" code block to add more description to the final commit message.
For more info, check https://github.com/tikv/pd/blob/master/CONTRIBUTING.md#format-of-the-commit-message.

-->

```commit-message
```
1. cancel timeout in loading function
2. use larger batch when loading from etcd
3. reduce times that call `buildRangeList` in label watcher

### Check List

<!-- Remove the items that are not applicable. -->

Tests

<!-- At least one of these tests must be included. -->

- Unit test
```
goos: linux
goarch: amd64
pkg: github.com/tikv/pd/pkg/mcs/scheduling/server/rule
cpu: AMD Ryzen 7 7840S with Radeon 780M Graphics         
```
| BenchmarkLoadLargeRules-16      |  | ||
| :---        |    ----: |        ----:  | ----: | 
| Master(timeout1h)     |  270616573162 ns/op        | 199973994432 B/op     |270541132 allocs/op|
| batch | 303262574 ns/op|	135643794 B/op	| 1146309 allocs/op|
| batch with larger limit (This PR) |208206918 ns/op |       128028446 B/op       |  1197805 allocs/op|
| This PR with easyjson | 201611560 ns/op       | 124696554 B/op    |     1132302 allocs/op|

#### cpu time

Master:
![image](https://github.com/tikv/pd/assets/19542290/463a11ce-4b9d-46ed-bd0a-445ff4f52b59)

This PR:
![image](https://github.com/tikv/pd/assets/19542290/2e3ce742-9236-4f95-b18c-46f510e2e85b)


### Release note

<!--

A bugfix or a new feature needs a release note. If there is no need to give a release note, just leave it with the `None`.

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

-->

```release-note
None.
```